+++ optuna/importance/_ped_anova/__init__.py
@@ -1,4 +1,4 @@
-from optuna.importance._ped_anova._evaluator import PedAnovaImportanceEvaluator
+from optuna.importance._ped_anova.evaluator import PedAnovaImportanceEvaluator
 
 
 __all__ = ["PedAnovaImportanceEvaluator"]
